requirements.txt
fastapi
uvicorn
psutil
structlog

# OpenTelemetry for distributed tracing (Stage 2)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib3
